@msbutler 2024-11-25
I'm also skeptical this would add more test coverage, given the additional monetary cost this would add. Like, we already run these benchmarks once a night, right? If the owning team gets a failure the morning after they merged a problematic patch, they should be able to quickly bisect the regression.