@sambhav-jain-16 2024-11-25
> Any chance we could narrow the set of tests in the required job a bit and/or split it across jobs such that we aren't regressing on end-to-end time for ci?

The script uses a single `BenchmarkKV` from the sql package [ref],(https://github.com/cockroachdb/cockroach/blob/eac4efe9be039d149aa8539362a1fd241c0b06d2/build/github/cockroach-microbench-ci.sh#L49) there's no way to split it since it is a single microbenchmark.

> If a patch changes the backupccl package, i don't see the need to run the microbenchmarks in the storage package. Before merging any sort of microbenchmark test as required in ci, can we limit the benchmarks run to the to packages that were changed in the commit?
I'm quite worried this change would lead to very flaky CI.

Same as above, since `BenchmarkKV` covers end-to-end benchmark for the in-memory cockroach, so we can't limit per package.

As per the flakiness part, this has been addressed by analysing the variance of the subsequent runs which were plotted [here](https://lookerstudio.google.com/reporting/d707c4cc-5b86-490f-891f-b023766bb89f/page/n1tCE). There were also changes by @nvanbenschoten to reduce the variance https://github.com/cockroachdb/cockroach/pull/131711.

Apart from the above changes, I have analysed past 1 month of microbenchmark and @srosenberg and me arrived at a threshold value of `20%` for a regression, this change was merged [here](https://github.com/cockroachdb/cockroach/pull/134810)

> We can't be making CI any longer to complete.

This can be addressed by reducing the no of count of the benchmark run but that will result in more variance in the benchmark runs. @srosenberg wdyt about this? If that's the trade-off we want to take for faster ci.

> Like, we already run these benchmarks once a night, right? If the owning team gets a failure the morning after they merged a problematic patch, they should be able to quickly bisect the regression.

https://github.com/cockroachdb/cockroach/issues/106661 mentions how roachperf is not been able to detect small regressions that have been incrementally added by PRs, the ci job aims to address this by stopping the PR from getting merge itself.