@rickystewart 2024-11-25
> there's no way to split it since it is a single microbenchmark.

Obviously, since `BenchmarkKV` is nothing more than a loop with 4 iterations, you could extract it into 4 different benchmarks.