@rickystewart 2024-11-25
> Apart from the above changes, I have analysed past 1 month of microbenchmark and @srosenberg and me arrived at a threshold value of 20% for a regression, this change was merged https://github.com/cockroachdb/cockroach/pull/134810

How many spurious failures have occurred since this change was made? Have any legitimate failures been observed?