@dt 2024-11-25
> https://github.com/cockroachdb/cockroach/issues/106661 mentions how roachperf is not been able to detect small regressions that have been incrementally added by PRs, the ci job aims to address this by stopping the PR from getting merge itself.

I'm not quite following this? If small regressions are hard to notice due to being small, how does looking for them before vs after merge change the scale of the regression we can reliably detect? It seems like f we're able to mechanically determine if one SHA is a regression compared to another SHA with high confidence -- such that we can run it on the PR's SHA vs its mergebase -- then it seems like we, could also run this once or twice a day comparing to the prior day's HEAD SHA to file an issue (with the delta commit log) too? Does this really need to be in the PR path?