@dt 2024-11-25
> forces the engineer to fix regressions in the PR itself

Do we need to _force_ engineers to do this? Is there evidence that if we present engineers with _reliable_ information that they just merged a significant regression that they won't act quickly to fix/revert/etc?

> waiting for the nightlies to run ... turn around time

While that's true, it isn't clear that that is _too much_ turn around time, or that cutting that down is worth the cost (either in $ or more importantly in productivity due to a slower CI pipeline). That said, if we really wanted something quicker I think we still have options that aren't to be in the blocking path of every PR, e.g. I believe we already have a CI that runs every time master's SHA has changed ("merge to master" or "publish bleeding edge" or something? dev-inf/rel-eng would know the details) in which the list of candidate changes compared to last run would be relatively small (just what was in the bors batch).