@srosenberg 2024-11-25
> Do we need to force engineers to do this? Is there evidence that if we present engineers with reliable information that they just merged a significant regression that they won't act quickly to fix/revert/etc?

I guess I would rephrase the above differently. The operation assumption is that the specifically chosen microbenchmarks are fairly _stable_, and the asserted performance difference denotes a sizable regression, with high probability. If that's the case, then how is this different from forcing engineers that the unit tests pass? Catching a regression after an offending PR is merged, translates to doing more (infra.) work. When would you not want to block PR, after learning that it contains a sizable performance regression?