@rickystewart 2024-11-26
OK, I ran the benchmark in question at `d0ff9c702fbf391c4a2e080b81a6c64400a39bad` and `8c34a77f5c6e806b359f20a6fd843817b1cc9d5e` ("after"). I could not reproduce any 20% regression introduced in `8c34a77...`. Can you reproduce this or does it indeed seem like this check is flaky?

```
before

BenchmarkKV/Delete/SQL/rows=100       	    1107	    975721 ns/op	  479261 B/op	    4988 allocs/op
BenchmarkKV/Delete/SQL/rows=1000      	     100	  16787909 ns/op	10634184 B/op	   56985 allocs/op
BenchmarkKV/Delete/SQL/rows=10000     	       7	 177139738 ns/op	142393733 B/op	  693768 allocs/op

BenchmarkKV/Delete/SQL/rows=100       	    1171	   1009993 ns/op	  486320 B/op	    4978 allocs/op
BenchmarkKV/Delete/SQL/rows=1000      	     100	  16331729 ns/op	10586844 B/op	   56972 allocs/op
BenchmarkKV/Delete/SQL/rows=10000     	       8	 172847656 ns/op	141608931 B/op	  668106 allocs/op

BenchmarkKV/Delete/SQL/rows=100       	     944	   1231771 ns/op	  488684 B/op	    5008 allocs/op
BenchmarkKV/Delete/SQL/rows=1000      	     100	  17260990 ns/op	10626778 B/op	   57001 allocs/op
BenchmarkKV/Delete/SQL/rows=10000     	       7	 178148024 ns/op	141316301 B/op	  680248 allocs/op

after
BenchmarkKV/Delete/SQL/rows=100       	    1096	    989900 ns/op	  480403 B/op	    4986 allocs/op
BenchmarkKV/Delete/SQL/rows=1000      	     100	  16811722 ns/op	10669168 B/op	   58635 allocs/op
BenchmarkKV/Delete/SQL/rows=10000     	       8	 186308281 ns/op	148710421 B/op	  681618 allocs/op

BenchmarkKV/Delete/SQL/rows=100       	    1086	   1036353 ns/op	  481556 B/op	    4987 allocs/op
BenchmarkKV/Delete/SQL/rows=1000      	     100	  16586513 ns/op	10619645 B/op	   57931 allocs/op
BenchmarkKV/Delete/SQL/rows=10000     	       7	 180607661 ns/op	143315194 B/op	  695340 allocs/op

BenchmarkKV/Delete/SQL/rows=100       	    1038	    983737 ns/op	  487190 B/op	    4993 allocs/op
BenchmarkKV/Delete/SQL/rows=1000      	     100	  16540953 ns/op	10630585 B/op	   57038 allocs/op
BenchmarkKV/Delete/SQL/rows=10000     	       8	 181020609 ns/op	146163115 B/op	  670403 allocs/op
```